feat✨[list-builder]: Add TableTreeComponent and modify access modifier
Added a new component, TableTreeComponent, to the list-builder module. This component includes features for tree structure data representation with drag-drop functionality. Also modified the access modifier of ChangeDetectorRef in BaseTableDefaultComponent from private to public.

This commit also introduces changes in the list-builder.module.ts file where necessary imports for MatTreeModule and CdkDragDrop are added along with declarations and exports for the newly created TableTreeComponent.