refactor: :hammer: Update condition checks and use viewController in components
In the CreateWizardComponent, BaseInputRelationDropdownComponent, and BaseTableDefaultComponent, the code has been refactored to use a viewController. This is either obtained from META.getViewController(model) or defaults to the model itself.

In CreateWizardComponent, an additional check for fileFields.length > 0 was added to ensure that there are elements in fileFields before proceeding.

In BaseInputRelationDropdownComponent, settings are now derived from viewController.relations instead of directly from getModelDefinition().relations.

In BaseTableDefaultComponent, listFields() and listActions() are now called on viewController instead of directly on the model.